# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NavPlat API
  description: Navigation Platform API Documentation - Complete API specification for Postman and other API clients
  version: 1.0.0
  contact:
    name: NavPlat Team
    email: support@navplat.local
servers:
  - url: http://localhost:5000/api
    description: Local Development Gateway
  - url: https://api.navplat.local
    description: Production Gateway
...
paths:
...
  /journeys:
    get:
      tags: [Journeys]
      summary: Get user's journeys
      description: Returns paginated list of journeys for the authenticated user (includes own journeys and shared journeys)
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: startDateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Filter journeys starting from this date
        - name: startDateTo
          in: query
          schema:
            type: string
            format: date-time
          description: Filter journeys starting until this date
      responses:
        '200':
          description: List of journeys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedJourneys'
        '401':
          description: Unauthorized
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    post:
      tags: [Journeys]
      summary: Create a new journey
      description: Creates a new journey for the authenticated user. Publishes JourneyCreated domain event.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJourneyRequest'
            example:
              startLocation: "Tirana, AL"
              startTime: "2025-04-11T08:15:00Z"
              arrivalLocation: "Durrës, AL"
              arrivalTime: "2025-04-11T09:05:00Z"
              transportType: "Commercial"
              distanceKm: 38.40
      responses:
        '201':
          description: Journey created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateJourneyResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
  /journeys/{id}:
    get:
      tags: [Journeys]
      summary: Get journey by ID
      description: Returns journey details. User must own the journey, have it shared with them, or be an Admin.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Journey ID
      responses:
        '200':
          description: Journey details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Journey'
        '404':
          description: Journey not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden - user doesn't have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized

    put:
      tags: [Journeys]
      summary: Update journey
      description: Updates an existing journey. Only the owner can update. Publishes JourneyUpdated domain event.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJourneyRequest'
      responses:
        '204':
          description: Journey updated successfully
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Journey not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden - user is not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized

    delete:
      tags: [Journeys]
      summary: Delete journey
      description: Deletes a journey. Only the owner can delete. Publishes JourneyDeleted domain event.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Journey deleted successfully
        '404':
          description: Journey not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden - user is not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: .NavPlat.Auth
      description: HTTP-only cookie containing authentication token
  schemas:
...
    Journey:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          example: "user-123"
        startLocation:
          type: string
          example: "Tirana, AL"
        startTime:
          type: string
          format: date-time
          example: "2025-04-11T08:15:00Z"
        arrivalLocation:
          type: string
          example: "Durrës, AL"
        arrivalTime:
          type: string
          format: date-time
          example: "2025-04-11T09:05:00Z"
        transportType:
          type: string
          enum: [Commercial, Cargo, Private, Charter, Military, Helicopter, Other]
          example: "Commercial"
        distanceKm:
          type: number
          format: decimal
          example: 38.40
        isFavorite:
          type: boolean
          example: false
        isShared:
          type: boolean
          example: false
        isDailyGoalAchieved:
          type: boolean
          example: false
    CreateJourneyRequest:
      type: object
      required: [startLocation, startTime, arrivalLocation, arrivalTime, transportType, distanceKm]
      properties:
        startLocation:
          type: string
          minLength: 1
          example: "Tirana, AL"
        startTime:
          type: string
          format: date-time
          example: "2025-04-11T08:15:00Z"
        arrivalLocation:
          type: string
          minLength: 1
          example: "Durrës, AL"
        arrivalTime:
          type: string
          format: date-time
          example: "2025-04-11T09:05:00Z"
        transportType:
          type: string
          enum: [Commercial, Cargo, Private, Charter, Military, Helicopter, Other]
          example: "Commercial"
        distanceKm:
          type: number
          format: decimal
          minimum: 0
          example: 38.40
    UpdateJourneyRequest:
      type: object
      required: [startLocation, startTime, arrivalLocation, arrivalTime, transportType, distanceKm]
      properties:
        startLocation:
          type: string
          minLength: 1
          example: "Tirana, AL"
        startTime:
          type: string
          format: date-time
          example: "2025-04-11T08:15:00Z"
        arrivalLocation:
          type: string
          minLength: 1
          example: "Durrës, AL"
        arrivalTime:
          type: string
          format: date-time
          example: "2025-04-11T09:05:00Z"
        transportType:
          type: string
          enum: [Commercial, Cargo, Private, Charter, Military, Helicopter, Other]
          example: "Commercial"
        distanceKm:
          type: number
          format: decimal
          minimum: 0
          example: 38.40
    CreateJourneyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
...
    PagedJourneys:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Journey'
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalCount:
          type: integer
          example: 150
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: "https://tools.ietf.org/html/rfc7231#section-6.5.1"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "Validation failed"
        instance:
          type: string
          format: uri
          example: "/api/journeys"
        traceId:
          type: string
          example: "0HN1234567890"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Validation errors by field name